激进的删除了学习之书

+++ src/main/java/net/minecraft/AgeOfMinecraft/entity/tame/EntityAbstractIllagers.java
@@ -24,7 +24,7 @@ public abstract class EntityAbstractIllagers extends EntityTameBase {
   
   @SideOnly(Side.CLIENT)
   protected boolean isAggressive(int mask) {
-    int i = (Byte) this.dataManager.get(AGGRESSIVE);
+    int i = this.dataManager.get(AGGRESSIVE);
     return ((i & mask) != 0);
   }
